refactor(nfs): split the state manager god file by concern
